refactor: create new component for LegIconWithA11y
