+++ lib/components/narrative/line-itin/connected-itinerary-body.js
@@ -1,7 +1,6 @@
 /* eslint-disable react/prop-types */
 // TODO: Typescript (otp-rr config object)
 import { connect } from 'react-redux'
-import { isTransit } from '@opentripplanner/core-utils/lib/itinerary'
 import {
   LegDescriptionHeadsignPrefix,
   PlaceName as PlaceNameWrapper,
@@ -24,15 +23,14 @@ import {
   getFirstLegStartTime,
   getLastLegEndTime
 } from '../../../util/itinerary'
-import { getFormattedMode } from '../../../util/i18n'
 import { injectIntl } from 'react-intl'
 import { MainPanelContent } from '../../../actions/ui-constants'
 import { setLegDiagram, setMapillaryId } from '../../../actions/map'
 import { setMainPanelContent, setViewedTrip } from '../../../actions/ui'
-import InvisibleA11yLabel from '../../util/invisible-a11y-label'
 import TripDetails from '../connected-trip-details'
 import TripTools from '../trip-tools'
 
+import LegIconWithA11y from './LegIconWithA11y'
 import RealtimeTimeColumn from './realtime-time-column'
 import TransitLegSubheader from './connected-transit-leg-subheader'
 
@@ -77,7 +75,6 @@ class ConnectedItineraryBody extends Component {
       activeItineraryTimeIndex,
       config,
       diagramVisible,
-      intl,
       itinerary,
       RouteDescriptionOverride,
       setActiveLeg,
@@ -86,7 +83,6 @@ class ConnectedItineraryBody extends Component {
       setMapillaryId,
       setViewedTrip
     } = this.props
-    const { LegIcon } = this.context
     const clonedItinerary = clone(itinerary)
 
     // Support OTP1 flex messages in Trip Details
@@ -129,18 +125,6 @@ class ConnectedItineraryBody extends Component {
           daysPrior: 0
         }
       }
-    }
-
-    const LegIconWithA11y = (props) => {
-      const { leg } = props
-      const { mode } = leg
-      const ariaLabel = isTransit(mode) ? getFormattedMode(mode, intl) : null
-      return (
-        <>
-          <LegIcon {...props} />
-          {ariaLabel && <InvisibleA11yLabel>{ariaLabel}</InvisibleA11yLabel>}
-        </>
-      )
     }
 
     return (
